+++ build.gradle
@@ -103,5 +103,5 @@ jdks {
 
 gradleTestUtils {
     gradleVersions = ['8.14.3']
-    configurationCacheEnabled = true
+    configurationCacheEnabled = false
 }
